Adding position abstraction

+++ dash/include/dash/Team.h
@@ -450,7 +450,7 @@ public:
     while (t && level > 0 && !(t->is_leaf())) {
       t = t->_child;
       level--;
-      DASH_ASSERT_MSG((int)t, "node is neither leaf nor has child");
+      DASH_ASSERT_MSG((long)t, "node is neither leaf nor has child");
     }
     return *t;
   }
@@ -460,7 +460,7 @@ public:
     Team *t = this;
     while (t && !(t->is_leaf())) {
       t = t->_child;
-      DASH_ASSERT_MSG((int)t, "node is neither leaf nor has child");
+      DASH_ASSERT_MSG((long)t, "node is neither leaf nor has child");
     }
     return *t;
   }
